Update the low multiplication code in fmpz_poly
Renamed parameters in the various low multiplication
routines from "trunc" to "n", to be in line with the
"_n" suffix and to provide for a consistent naming
convention with the high and middle products, to which
the word "trunc" does not apply.

Remove one use of "memcpy" and use a simple for loop
instead.