Initial inclusion of the semi-automated manual.
The file doc.c contains code of a very simple parser (implemented
using a finite state machine) for the current documentation files
in $module/doc/$module.  There are six methods near the top of the
code which print output whenever a new group (addition,
multiplication, etc), a new function signature, or a new function
description is opened or closed.

The file doc.tex will eventually contain the actual source code
of the manual, together with some "input" macros to include the
LaTeX files generated using the above program.

The "Makefile" currently provides the options to call "make doc"
and "make clean", the first of which should compile doc.c, then
call the resulting executable and finally latex to produce
doc.dvi.  This could potentially fail if some packages are missing
in the local LaTeX set-up.