Modifies fmpz_randtest().
This patch sets a higher priority to special absolute values such
as 0, 1, COEFF_MAX and LONG_MAX.

This patch also fixes a bug (which did not seem to have surfaced
so far).  In the case of large integers, there was no _fmpz_demote_val()
yet.  This probably did not show so far because the top bits in the
integers returned by mpz_rrandomb() are likely to be set.  But this
is not documented in the GMP manual.