Add a function "fmpq_poly_debug" to print debug information.
This function does not appear in the documentation at this point.
In its current form, it will do to find and fix the bugs that
show up right now.  But I'll use it for a while to see whether
this is exactly what I want.  Then, perhaps, I'll include it in
the documentation.