Fix "control reaches end of non-void function" in fmpq_poly/debug
Sets it to always return 1.  Of course, we could change the
signature to void rather than int.  But perhaps at some point
one might want to be a little stricter and check for return
values of printf (or maybe move this to fprintf where errors
are much more likely than for printf) and always returning 1
now would make it easier to change this later.