Remove "fmpz_remove".
Sebastian initially added this function to the fmpz module to
be used from the new padic module.  In the meantime, there is
a more specific version in use in the padic module and the
function fmpz_remove is obsolete.  In the current state, it
wasn't anything but a very naive wrapper around "mpz_remove"
anyway, hence it can be removed.