Rewrite nmod_poly_xgcd_euclidean.
The new implementation follows Cohen's book, unrolling
the first loop iteration.  It is the same implementation
as in fmpz_mod_poly.  Timings for the test code in
nmod_poly/test/t-xgcd_euclidean improved from .78s to .56s
so I think this is a positive change not just from the
code flow (no more "steps" to take care of the conditional
allocation) but also in terms of performance.  This is
achieved because we only take care of one cofactor during
the loop iteration and compute the second one at the very
end.