Make config adjust headers based on --reentrant.
We shouldn't make the FLINT_REENTRANT flag a user responsibility.  Currently,
users of Flint must use -DFLINT_REENTRANT (or not) based on how the library was
built.  While there is no ideal solution - users can always copy a binary
between machines with dynamic Flint libraries using differing compilation
parameters - an improvement on the current state is to make the header files
match the library which which it was installed.