configure: fix libflint install_name on Mac OS X
Before this change, the libflint install_name on OS X was
  libflint-$FLINT_MAJOR.$FLINT_MINOR.$FLINT_PATCH.dylib

But that means that any change to the minor version, or even the patch level,
broke compatibility, as the dynamic linker would be looking for exactly
that filename. For the Linux users out there: the above would be akin to
using the soname
   libflint.so.$FLINT_MAJOR.$FLINT_MINOR.$FLINT_PATCH

The correct approach is to use the install_name
   libflint-$FLINT_MAJOR.dylib
just like the Linux build uses the soname
   libflint.so.$FLINT_MAJOR

The compatibility_version / current_version flags then take care of preventing
the user from strange errors due to downgrading to an older dylib version.