Remove NTL_CLIENT from NTL-interface.h
With a typical NTL installation, NTL_CLIENT expands to 'using namespace
std; using namespace NTL', or something like that. This breaks lots of
things when stuck in a C++ header.

I think the NTL_NNS macro should expand to 'NTL ::' in a standard NTL
installation, and the empty string in a "no namespace" installation of
NTL, so that's what I've used here. Works for me, at least.