Undefine count_leading_zeros and count_trailing_zeros in longlong.h; fixes #306
The fallbacks at the end of the file are only defined when there is no
definition yet. When building SageMath it can happen that the symbols
are already defined in a givaro header. That lead to build failures on
architectures that are supposed to use the fallback here, like arm64
and ppc64el.