build: rename HAVE_PTHREAD to FLINT_USES_PTHREAD
A library such as flint should always try to namespace its own
defines somehow, lest flint's defines could interfere or override a
macro of the same name in a project that happens to use HAVE_* macros
of its own.

HAVE_PTHREAD's value influences the ABI of e.g. fmpz_block_header_s,
so the macro does not just describe a system property at the time of
building, but also the state of a built-and-installed libflint, so to
that end, I renamed HAVE_PTHREAD to FLINT_USES_PTHREAD.