CMake/FindGMP.cmake: compile a test program to check the gmp version.
The existing GMP version check consists of grepping the contents of
the gmp.h header to parse out a few constants. This test fails, at
least, on Gentoo, where the usual header file is a wrapper that
includes the true header to allow for simultaneous 32/64-bit
support.

This commit updates FindGMP to compile a test program against gmp.h
that compares version constants using the C preprocessor. If GMP is
too old, the test program will fail to compile and CMake will know.