Fix Clang warnings and faulty init/alloc in test
fq_poly_templates/test/t-sqrt.c in a test had initialized a fq_nmod_t
while the ctx was uninitialized (randtest had not been set). Hence the
allocation size for it was extremely large. This was fixed.

Moreover, Clang had warnings about indentation (from mixing tabs
and spaces) and a warning about a combination of negation and inequality
of a template. These was also fixed.