Changed documentation to match headers.
Most of this change is cosmetic, geting code and documentation to match
const/mutable or signed/unsigned.  The change is almost always a change to
documentation. arb.h and fmpz_mod_mpoly.h were changed so that `set_ui`
functions were modified to take an ulong instead of a slong.

This does fix a fair number of typos, copy paste errors, etc., where the code
and the documentation diverge considerably.