CI: limit which branches run CI tests
Right now when people make PRs from branches which live in the main
repository `flintlib/flint` (as opposed to in a branch), the CI will run
all tests *twice*: once for the branch HEAD, and once for the pull
request HEAD (which is a merge commit created from merging the branch
HEAD into trunk). This is mostly useless, so let's disable that to
conserve more CI resources. Only run those CI tests on trunk and on
release branches.