Add assembly routines for mul and sqr
- flint_mpn_mul_M_N for M = 1, ..., 16 and N = 1, ..., min(M, 8)

- flint_mpn_sqr_M for M = 1, ..., 7

- Check during configuration whether the assembly routines are
  available. They are only available if M4 is available (just like in
  GMP) as we need to process the files differently for different
  systems.

NOTE: It would be nice to support bigger squaring, but that seems hard
at the moment. The problem is that we have to push to the stack, which
is slow compared to just using registers. There is a functional squaring
routine for M = 8, but that is only five percent faster than
mpn_sqr_basecase.