do not use find_library() in flint_ctypes
Using find_library(), it seems impossible to use LD_LIBRARY_PATH to load
a specific version of libflint when there is one the system library
search path. On my system at least, however, specifying the name alone
seems to use the normal ld.so/dlopen() search algorithm and works both
for an installed library and one in LD_LIBRARY_PATH.

(According to the Python docs, find_library() tries to locate the shared
library in a platform-dependent way vaguely similar to the way it is
done at _compile_ time. On Linux, it appears to use the ldconfig cache.)