polynomial dispersion and shiftless decomposition
Both implementations start from an irreducible factorization: the
dispersion follows Man and Wright (1994) and the shiftless decomposition
is completely naive.

This should work over FMPZ and QQBAR, and, in favorable cases, over
complex Calcium fields and CC_ACB. For fields like FMPQ or real Calcium
fields, I think all that is missing is an implementation of
GR_METHOD_POLY_FACTOR.

Note that Gerhard, Giesbrecht, Storjohann, and Zima (2003) give a
shiftless decomposition algorithm that only needs polynomial gcds once
the dispersion set of the input is known, and a dispersion algorithm
that only requires a squarefree factorization and low-precision p-adic
factorization.