generic sorting (and related) functions
It is a bit difficult to reliably detect if the system provides a
qsort_r function with the required behavior, so for now I am only using
using it with glibc. In principle we could also accept POSIX.1-2024 and
add cases for BSD-style qsort_r as well as qsort_s.