Add V Makefile variable for verbose builds
make V=0 (the default) prints "CC filename.c"

make V=1 prints the full gcc command with build flags, which is useful
for parsing build logs to see if particular flags are being used.
For example, Debian does this to check if hardening flags are used
when building packages using the program blhc.