problem with negative leading term
in Hecke:
julia> K, a = number_field(polynomial(QQ, [-1, 0, -16]))
(Number field of degree 2 over QQ, _a)

julia> Hecke.norm_div(8*a-56, ZZ(1), 100)

would compute the resultant, and then divide by a power of the leading
coefficient using _fmpq_mul. But the last argument, the power of the
leading coeff. is negative and that violates flints rationals.